+++ CHANGES.rst
@@ -862,6 +862,7 @@ astropy.wcs
 
 Bug fixes
 ---------
+  
 
 astropy.config
 ^^^^^^^^^^^^^^
@@ -968,6 +969,8 @@ astropy.modeling
 - Removed a limitation of fitting of data with units with compound models
   without units when the expression involves operators other than addition
   and subtraction. [#10415]
+
+- Fixed a problem with fitting ``Linear1D`` and ``Planar2D`` ain model sets. [#10623]
 
 astropy.nddata
 ^^^^^^^^^^^^^^